+++ src/app/search/page.tsx
@@ -1,28 +1,14 @@
-'use client'
-
-import { useEffect, useState } from 'react'
+import GoToCart from '@/components/AppBar/GoToCart'
 import NonSearch from '@/containers/search/NonSearch'
-// import OnSearchList from '@/containers/search/OnSearchList'
 import SearchInputBar from '@/containers/search/SearchInputBar'
-// import { searchValueState } from '@/states/searchAtom'
-// import { useRecoilValue } from 'recoil'
 
 export default function SearchPage() {
-  // const searchValue = useRecoilValue(searchValueState)
-
-  const [mounted, setMounted] = useState(false)
-
-  useEffect(() => {
-    setMounted(true)
-  }, [])
-
   return (
-    mounted && (
-      <>
-        <SearchInputBar />
-        <NonSearch />
-        {/* {!searchValue ? <NonSearch /> : <OnSearchList />} */}
-      </>
-    )
+    <>
+      <SearchInputBar>
+        <GoToCart />
+      </SearchInputBar>
+      <NonSearch />
+    </>
   )
 }
