+++ src/components/MainNav/state.ts
@@ -23,26 +23,26 @@ export const NavItems: NavType[] = [
   {
     id: 4,
     title: '명품',
-    url: '/luxury',
+    url: '/category/%EB%AA%85%ED%92%88',
   },
   {
     id: 5,
     title: '뷰티',
-    url: '/beauty',
+    url: '/category/%EB%B7%B0%ED%8B%B0',
   },
   {
     id: 6,
     title: '패션',
-    url: '/fashion',
+    url: '/category/%ED%8C%A8%EC%85%98%EC%9D%98%EB%A5%98',
   },
   {
     id: 7,
     title: 'SSG.TV',
-    url: '/ssg-tv',
+    url: '',
   },
   {
     id: 8,
     title: '브랜드',
-    url: '/brand',
+    url: '',
   },
 ]
